DetailScrollView Class를 만들어서 코드 분리
